moved screen calibration to screenspace.py from monopoly and player
