+++ player.py
@@ -5,9 +5,6 @@ import style as s
 from style import COLORS
 import screenspace as ss
 import modules as m
-import platform
-import ctypes
-import shutil
 import networking as n
 
 game_running = False
@@ -287,81 +284,17 @@ def gat() -> int:
     """
     return active_terminal
 
-def make_fullscreen():
-    current_os = platform.system()
-
-    if current_os == "Windows":
-        # Maximize terminal on Windows
-        user32 = ctypes.WinDLL("user32")
-        SW_MAXIMIZE = 3
-        hWnd = user32.GetForegroundWindow()
-        user32.ShowWindow(hWnd, SW_MAXIMIZE)
-
-    elif current_os == "Linux" or current_os == "Darwin":
-        # Maximize terminal on Linux/macOS
-        os.system("printf '\033[9;1t'")
-    else:
-        print(f"Fullscreen not supported for OS: {current_os}")
-    
-def print_with_wrap(char, start_row, start_col):
-    # Get the terminal size
-    terminal_size = shutil.get_terminal_size()
-    width = terminal_size.columns
-    
-    # If the position exceeds the terminal width, handle wrapping
-    if start_col >= width:
-        # Calculate new row and column if it exceeds width
-        new_row = start_row + (start_col // width)
-        new_col = start_col % width
-        print(f"\033[{new_row};{new_col}H" + char, end="")
-    else:
-        # Default print
-        print(f"\033[{start_row};{start_col}H" + char, end="")
-
-def scaling_print():
-    os.system('cls' if os.name == 'nt' else 'clear')
-    current_os = platform.system()
-    if current_os == "Darwin":
-        # Print out instructions for macOS users
-        print("Please use Ctrl + \"Command\" + \"+\" or Ctrl + \"Command\" + \"-\" to zoom in/out and ensure everything is visible. Press enter to continue to scaling.")
-    else:
-        # Print out instructions for Linux/Windows users
-        print("Please use \"Ctrl\" + \"-\" or \"Ctrl\" + \"+\" to zoom in/out and ensure everything is visible. Press enter to continue to scaling screen.")
-    print("Proper scaling should only displays 4 cross that marks the corners of the board.")
-    print("If you are having trouble with scaling, try entering r to reset the display.")
-    print("After finishing scaling, please press enter to continue.")
-    scaling_test = input()
-
-    os.system('cls' if os.name == 'nt' else 'clear')
-    ss.print_screen()
-
-    print_with_wrap("X", 0, 0)
-    print_with_wrap("X", 0, 153)
-    print_with_wrap("X", 43, 153)
-    print_with_wrap("X", 43, 0)
-    print(f"\033[44;0H" + "Press enter to play or enter r to reset the display.", end="")
-    scaling_test = input()
-    while scaling_test != "":
-        os.system('cls' if os.name == 'nt' else 'clear')
-        ss.print_screen()
-        print_with_wrap("X", 0, 0)
-        print_with_wrap("X", 0, 153)
-        print_with_wrap("X", 43, 153)
-        print_with_wrap("X", 43, 0)
-        print(f"\033[44;0H" + "Press enter to play or enter r to reset the display.", end="")
-        scaling_test = input()
-    os.system('cls' if os.name == 'nt' else 'clear')
 
 if __name__ == "__main__":
-    make_fullscreen()
     """
     Main driver function for player.
     """
     get_graphics()
 
-    initialize()
+    # initialize()
 
-    scaling_print()
+    ss.make_fullscreen()
+    ss.calibrate_screen('player')
 
     # Prints help in quadrant 2 to orient player.
     os.system("cls")
@@ -370,7 +303,6 @@ if __name__ == "__main__":
     ss.update_quadrant_2(2, data=None)
     ss.update_quadrant_2(3, data=None)
     ss.update_quadrant_2(4, data=None)
-    ss.update_active_terminal(1)
     get_input()
 
     # ss.print_board(text_dict.get('gameboard'))
